Continuing the refactor
I technically took away functionality, but I wanted to check in my "compiling" and running version of Circuitent.

I abstracted out an Editor interface to make the common stuff that will be shared between the various editor views. I also started to add functionality to allow the CircuitentDrawingSurface to be scrollable. It sort of works, but the parent window seems to want to resize initially to the size of the CDS's initial size. No idea why...GTK weirdness I'm sure.